refactor: renamed custom hooks
