+++ hooks/useFileUpload.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useToast } from "@/hooks/use-toast";
+import { useToast } from "@/hooks/useToast";
 
 export function useFileUpload() {
   const { toast } = useToast();
